Move Connection to separate module
